+++ src/common/mod.rs
@@ -6,10 +6,10 @@ pub mod gas;
 pub mod dive_segment;
 pub mod otu;
 
-pub fn bar_mtr(bar: f32) -> f32 {
+pub fn bar_mtr(bar: f64) -> f64 {
     (bar-1.0) * 10.0
 }
 
-pub fn mtr_bar(mtr: f32) -> f32 {
+pub fn mtr_bar(mtr: f64) -> f64 {
     (mtr/10.0) + 1.0
 }
